requirements.txt
flask
flask-cors
psycopg2-binary
pandas
plotly
streamlit
requests
yfinance
cryptography
python-dotenv
bcrypt
